requirements.txt
alembic==1.16.4
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
MouseInfo==0.1.3
mysql-connector-python==9.3.0
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.1
pytweening==1.2.0
requests==2.32.4
SQLAlchemy==2.0.42
typing_extensions==4.14.1
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1